Fix numeric type mixing in energy calculations

+++ app/src/main/java/com/example/socialbatterymanager/features/energy/EnergyManager.kt
@@ -121,24 +121,24 @@ class EnergyManager {
                     id = 1,
                     title = "Team Meeting",
                     description = "Weekly standup with development team",
-                    startTime = todayStart + (9 * 60 * 60 * 1000), // 9:00 AM
-                    endTime = todayStart + (10.5 * 60 * 60 * 1000), // 10:30 AM
+                    startTime = todayStart + (9 * 60 * 60 * 1000L), // 9:00 AM
+                    endTime = todayStart + (10.5 * 60 * 60 * 1000).toLong(), // 10:30 AM
                     source = "google",
                 ),
                 CalendarEvent(
                     id = 2,
                     title = "Lunch with Sarah",
                     description = "Catching up with close friend",
-                    startTime = todayStart + (12 * 60 * 60 * 1000), // 12:00 PM
-                    endTime = todayStart + (13 * 60 * 60 * 1000), // 1:00 PM
+                    startTime = todayStart + (12 * 60 * 60 * 1000L), // 12:00 PM
+                    endTime = todayStart + (13 * 60 * 60 * 1000L), // 1:00 PM
                     source = "manual",
                 ),
                 CalendarEvent(
                     id = 3,
                     title = "Client Presentation",
                     description = "Important project presentation",
-                    startTime = todayStart + (15 * 60 * 60 * 1000), // 3:00 PM
-                    endTime = todayStart + (16.5 * 60 * 60 * 1000), // 4:30 PM
+                    startTime = todayStart + (15 * 60 * 60 * 1000L), // 3:00 PM
+                    endTime = todayStart + (16.5 * 60 * 60 * 1000).toLong(), // 4:30 PM
                     source = "outlook",
                 )
             )
